Added a hotdog button

+++ src/main/java/ca/bcit/comp2522/termproject/comp2522202310termproject360/Item.java
@@ -2,6 +2,6 @@ package ca.bcit.comp2522.termproject.comp2522202310termproject360;
 
 public interface Item {
 
-    int getPassiveIncome();
+    double getPassiveIncome();
 
 }
